+++ inst/tarpuy/www/auth.R
@@ -1,4 +1,6 @@
 library(googleAuthR)
+library(gargle)
+library(httr)
 
 # check loaded package
 check_package_loaded <- function(package_name){
@@ -27,14 +29,14 @@ gar_js_getToken <- function(token,
                             client.id     = getOption("googleAuthR.webapp.client_id"),
                             client.secret = getOption("googleAuthR.webapp.client_secret")){
   check_package_loaded("shiny")
-  gar_app <- oauth_app("google", key = client.id, secret = client.secret)
+  gar_app <- httr::oauth_app("google", key = client.id, secret = client.secret)
   
   scope_list <- getOption("googleAuthR.scopes.selected")
   
   # Create a Token2.0 object consistent with the token obtained from gar_auth()
   token_formatted <-
     Token2.0$new(app = gar_app,
-                 endpoint = oauth_endpoints("google"),
+                 endpoint = httr::oauth_endpoints("google"),
                  credentials = list(access_token = token$access_token,
                                     token_type = token$token_type,
                                     expires_in = token$expires_in,